// frontend/src/app/core/common/enum/description-template-field-type.ts
export enum DescriptionTemplateFieldType {
	SELECT = "select",
	BOOLEAN_DECISION = "booleanDecision",
	RADIO_BOX = "radiobox",
	INTERNAL_ENTRIES_PLANS = "internalEntitiesPlans",
	INTERNAL_ENTRIES_DESCRIPTIONS = "internalEntitiesDescriptions",
	CHECK_BOX = "checkBox",
	FREE_TEXT = "freetext",
	TEXT_AREA = "textarea",
	RICH_TEXT_AREA = "richTextarea",
	UPLOAD = "upload",
	DATE_PICKER = "datePicker",
	TAGS = "tags",
	REFERENCE_TYPES = "referenceTypes",
	DATASET_IDENTIFIER = "datasetIdentifier",
	VALIDATION = "validation"
}